requirements.txt
sanic==20.12.0
python-decouple==3.1
sanic-openapi==0.6.2
squareup==8.0.0.20201216
